Pill renaming classes to fit dart convention
