+++ lib/widget/pill_widget.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:pill/bloc/pill_event.dart';
 import 'package:pill/bloc/pill_bloc.dart';
-import 'package:pill/model/PillToTake.dart';
-import 'package:pill/service/DateService.dart';
+import 'package:pill/model/pill_to_take.dart';
+import 'package:pill/service/date_service.dart';
 
 class PillWidget extends StatelessWidget {
